feature(#8): Move ITs to separate test package
