+++ app/src/test-integration/kotlin/ch/fitfusion/backfusion/app/TestUtils.kt
@@ -1,4 +1,4 @@
-package integration.test
+package ch.fitfusion.backfusion.app
 
 import io.restassured.specification.RequestSpecification
