+++ app/src/test/kotlin/ch/fitfusion/backfusion/app/BackFusionApplicationKtTest.kt
@@ -1,12 +1,9 @@
 package ch.fitfusion.backfusion.app
 
 import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
 import kotlin.test.assertTrue
 
-@ActiveProfiles("dev")
-@SpringBootTest
+@BackFusionApplicationTest
 class BackFusionApplicationTests {
 
     @Test
