+++ pom.xml
@@ -32,6 +32,8 @@
         <kotlin.code.style>official</kotlin.code.style>
         <kotlin.compiler.jvmTarget>1.17</kotlin.compiler.jvmTarget>
         <kotlin.version>1.9.10</kotlin.version>
+        <surefire-plugin.version>3.2.1</surefire-plugin.version>
+        <failsave-plugin.version>2.22.2</failsave-plugin.version>
     </properties>
 
     <dependencyManagement>
@@ -220,8 +222,24 @@
                 </plugin>
 
                 <plugin>
-                    <artifactId>maven-failsafe-plugin</artifactId>
-                    <version>2.22.2</version>
+                    <groupId>org.codehaus.mojo</groupId>
+                    <artifactId>build-helper-maven-plugin</artifactId>
+                    <version>3.1.0</version>
+
+                    <executions>
+                        <execution>
+                            <id>add-integration-test-source</id>
+                            <phase>generate-test-sources</phase>
+                            <goals>
+                                <goal>add-test-source</goal>
+                            </goals>
+                            <configuration>
+                                <sources>
+                                    <source>src/test-integration/kotlin</source>
+                                </sources>
+                            </configuration>
+                        </execution>
+                    </executions>
                 </plugin>
 
                 <plugin>
